Se refactorizo un poco el codigo.
